bookmarks/requirements.txt
asgiref==3.3.0
cffi==1.14.3
cryptography==3.2.1
Django==3.0.11
django-ckeditor==6.0.0
django-extensions==2.2.5
django-js-asset==1.2.2
easy-thumbnails==2.7
Pillow==7.0.0
pycparser==2.20
pyOpenSSL==19.0.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
Werkzeug==0.16.0
